mm3438: * Adding javadoc comments to interfaces and public objects. * making package protected all not public classes & methods. * moving MnuboSDKFactory to service package.
