+++ client/src/main/java/com/mnubo/java/sdk/client/services/AbstractSDKService.java
@@ -1,14 +1,14 @@
 /*
  * ---------------------------------------------------------------------------
- * 
+ *
  * COPYRIGHT (c) 2015 Mnubo Inc. All Rights Reserved.
- * 
+ *
  * The copyright to the computer program(s) herein is the property of Mnubo Inc. The program(s) may be used and/or
  * copied only with the written permission from Mnubo Inc. or in accordance with the terms and conditions stipulated in
  * the agreement/contract under which the program(s) have been supplied.
- * 
+ *
  * Author: marias Date : Aug 4, 2015
- * 
+ *
  * ---------------------------------------------------------------------------
  */
 
@@ -20,39 +20,34 @@ import org.springframework.web.client.RestTemplate;
 
 import com.mnubo.java.sdk.client.config.MnuboSDKConfig;
 
-public abstract class AbstractSDKService
-{
+abstract class AbstractSDKService {
 
     private final RestTemplate template;
     private final CredentialHandler credential;
     private final MnuboSDKConfig config;
 
-    public AbstractSDKService( RestTemplate template, CredentialHandler credential, MnuboSDKConfig config )
-    {
+    AbstractSDKService(RestTemplate template, CredentialHandler credential, MnuboSDKConfig config) {
         this.credential = credential;
         this.template = template;
         this.config = config;
     }
 
-    protected < T > T postRequest( String url, Class< T > objectClass, Object object )
-    {
+    protected <T> T postRequest(String url, Class<T> objectClass, Object object) {
         // entity
-        HttpEntity< ? > request = new HttpEntity< Object >( object , getAutorizationHeader() );
+        HttpEntity<?> request = new HttpEntity<Object>(object, getAutorizationHeader());
 
         // send request
-        return template.postForObject( url , request , objectClass );
+        return template.postForObject(url, request, objectClass);
     }
 
-    protected HttpHeaders getAutorizationHeader()
-    {
+    protected HttpHeaders getAutorizationHeader() {
         // header
         HttpHeaders headers = new HttpHeaders();
-        headers.set( "Authorization" , credential.getAutorizationToken() );
+        headers.set("Authorization", credential.getAutorizationToken());
         return headers;
     }
 
-    protected MnuboSDKConfig getConfig()
-    {
+    protected MnuboSDKConfig getConfig() {
         return config;
     }
