+++ client/src/main/java/com/mnubo/java/sdk/client/spi/OwnersSDK.java
@@ -1,14 +1,14 @@
 /*
  * ---------------------------------------------------------------------------
- * 
+ *
  * COPYRIGHT (c) 2015 Mnubo Inc. All Rights Reserved.
- * 
+ *
  * The copyright to the computer program(s) herein is the property of Mnubo Inc. The program(s) may be used and/or
  * copied only with the written permission from Mnubo Inc. or in accordance with the terms and conditions stipulated in
  * the agreement/contract under which the program(s) have been supplied.
- * 
+ *
  * Author: marias Date : Jul 27, 2015
- * 
+ *
  * ---------------------------------------------------------------------------
  */
 
@@ -16,8 +16,19 @@ package com.mnubo.java.sdk.client.spi;
 
 import com.mnubo.java.sdk.client.models.Owner;
 
-public interface OwnersSDK
-{
-    void create( Owner owner );
+/**
+ * Owner SDK Client. This interface gives access to handle Owners.
+ *
+ * @author Mauro Arias
+ * @since 2015/07/22
+ */
+public interface OwnersSDK {
+
+    /**
+     * Allows create owners.
+     *
+     * @param Owner, Owner bean to be created.
+     */
+    void create(Owner owner);
 
 }
